+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog of threedigrid-builder
 1.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Interpret non-finite raster values (NaN, Inf, -Inf) as nodata.
 
 
 1.1.0 (2022-01-24)